Add architecture constants to util.go so that they use the new SupportedArchs field of Function.
Adds X86 and ARM constants for supported architectures. Normally this field is assigned by the node who registers the function the first time, and then pushed to etcd so that it can be retrieved by other nodes. Here in the test the function registration is skipped/done manually through etcd directly, and so it was necessary to manually include this field and its values.