Implement architecture-aware offloading, with choice caching and choosing the (compatible) node with the most free memory.
This commit adds caching for offloading choices to improve performance. It now caches node selection for a period and uses available memory to pick a suitable node. It also considers only the nodes whose architecture is supported by the runtime of the function/request. E.g.: A function with a runtime that will only run on x86 will be offloaded to the neighbor with an x86 architecture (if there is any). If there is more than one, the one with the most memory available will be chosen. The choice will be cached for a pre-defined amount of time (this time can be fine-tuned in future commits)