Add architecture-aware load balancer
Implements an ArchitectureAwareBalancer that
distributes function invocations to nodes based on their
CPU architecture (ARM or x86) and function runtime support.
It uses two separate consistent hash rings for ARM and x86
nodes and includes logic for architecture selection and
tie-breaking (even tho the tie-break should be improved to leverage execution on specific architectures in the future).