Making consistent hashing memory-aware
Introduces a MemoryChecker interface and its default implementation to
enable memory-aware selection of nodes in the consistent hash ring.
This ensures that chosen nodes have sufficient memory for function
execution, falling back to other nodes in the ring if the primary
candidate lacks resources. The search for the fallback node is done by navigating the ring always in the same wat, to make sure that we leverage warm starts as much as possible in this scenario.

Introduced new file types.go to make the code more readable.

The memory checker is an interface with a default implementation to make future unit-tests easier and more reliable.