Improve info updates of load balancer from targets
Refactor load balancer to use HTTP requests for fetching node status. This ensures
that the `NodeMetrics` cache is periodically updated with the freshest information
from target nodes. It is also updated every time a target sends back a response to the ProxyBalancer.

Also, modify `GetServerStatus` in `api.go` to record the
`LastUpdateTime` for LB updates. This helps maintain more
accurate node metrics by providing a reliable timestamp for updates.