Add Go 1.25 runtime support
This runtime exposes a new package ("serverledge") that can be imported as a library to the user code. The user can then simply define their handler and call serverledge.Start(userHandler).
Serverledge expects (for the moment) a .zip file with both amd64 and arm64 executable. Since this runtime supports both architecture, it can actually be used even when specifying that only one architecture is supported with minimal changes needed to Serverledge's code.